+++ spec/stat_tracker_spec.rb
@@ -2,7 +2,7 @@ require 'simplecov'
 require 'pry'
 SimpleCov.start
 
-require './lib/common_stat'
+require './lib/game'
 require './lib/stat_tracker'
 
 RSpec.describe 'StatTracker' do
@@ -40,12 +40,14 @@ RSpec.describe 'StatTracker' do
   describe '#highest_total_score' do
     it 'returns the highest sum of the winning and losing teams scores' do
       expect(@stat_tracker.highest_total_score).to be_a Integer
+      expect(@stat_tracker.highest_total_score).to eq 11
     end 
   end
 
   describe '#lowest_total_score' do
     it 'returns the highest sum of the winning and losing teams scores' do
       expect(@stat_tracker.lowest_total_score).to be_a Integer
+      expect(@stat_tracker.lowest_total_score).to eq 0
     end 
   end
 
@@ -60,6 +62,21 @@ RSpec.describe 'StatTracker' do
     it 'percentage(float) of games that a home team has won (rounded to the nearest 100th)' do
       expect(@stat_tracker.percentage_visitor_wins).to be_a Float
       expect(@stat_tracker.percentage_visitor_wins).to eq 0.36
+    end
+  end
+
+  describe '#count_of_games_by_season' do
+    it 'returns a hash that counts the total games by season' do
+      expected = {
+        "20122013"=>806,
+        "20162017"=>1317,
+        "20142015"=>1319,
+        "20152016"=>1321,
+        "20132014"=>1323,
+        "20172018"=>1355
+      }
+      expect(@stat_tracker.count_of_games_by_season).to be_a Hash
+      expect(@stat_tracker.count_of_games_by_season).to eq expected
     end
   end
